+++ src/main/java/io/myzticbean/finditemaddon/scheduledtasks/Task15MinInterval.java
@@ -19,6 +19,7 @@
 package io.myzticbean.finditemaddon.scheduledtasks;
 
 import io.myzticbean.finditemaddon.FindItemAddOn;
+import io.myzticbean.finditemaddon.utils.async.VirtualThreadScheduler;
 import io.myzticbean.finditemaddon.utils.json.ShopSearchActivityStorageUtil;
 import io.myzticbean.finditemaddon.utils.warp.WarpUtils;
 import org.bukkit.Bukkit;
@@ -30,7 +31,8 @@ public class Task15MinInterval implements Runnable {
     @Override
     public void run() {
         // v2.0.6.0 - Changed tasks to run in async thread
-        Bukkit.getScheduler().runTaskAsynchronously(FindItemAddOn.getInstance(), () -> {
+        // v2.0.7.7 - Switched to virtual thread
+        VirtualThreadScheduler.runTaskAsync(() -> {
             WarpUtils.updateWarps();
             ShopSearchActivityStorageUtil.syncShops();
         });