+++ src/main/java/io/myzticbean/finditemaddon/utils/warp/WarpUtils.java
@@ -21,20 +21,20 @@ package io.myzticbean.finditemaddon.utils.warp;
 import io.myzticbean.finditemaddon.dependencies.EssentialsXPlugin;
 import io.myzticbean.finditemaddon.dependencies.PlayerWarpsPlugin;
 import io.myzticbean.finditemaddon.FindItemAddOn;
+import lombok.experimental.UtilityClass;
 
 /**
  * @author myzticbean
  */
+@UtilityClass
 public class WarpUtils {
     public static void updateWarps() {
         if(FindItemAddOn.getConfigProvider().shopGUIItemLoreHasKey("{NEAREST_WARP}")) {
-            if(FindItemAddOn.getConfigProvider().NEAREST_WARP_MODE == 1
-                    && EssentialsXPlugin.isEnabled()) {
+            if(FindItemAddOn.getConfigProvider().NEAREST_WARP_MODE == 1 && EssentialsXPlugin.isEnabled()) {
                 EssentialsXPlugin.updateAllWarps();
 
             }
-            else if(FindItemAddOn.getConfigProvider().NEAREST_WARP_MODE == 2
-                    && PlayerWarpsPlugin.getIsEnabled()) {
+            else if(FindItemAddOn.getConfigProvider().NEAREST_WARP_MODE == 2 && PlayerWarpsPlugin.getIsEnabled()) {
                 PlayerWarpsPlugin.updateAllWarpsFromAPI();
             }
         }